add handler cpu

+++ TaskAPI-2-1/MetricsAgent/DAL/TimeSpanHandler.cs
@@ -7,11 +7,12 @@ using System.Data;
 
 namespace MetricsAgent.DAL
 {
-    public class TimeSpanHandler: SqlMapper.TypeHandler<DateTimeOffset>
+    public class TimeSpanHandler : SqlMapper.TypeHandler<DateTimeOffset>
     {
         public override DateTimeOffset Parse(object value)
-                    => DateTimeOffset.FromUnixTimeSeconds((long)value);
-
+        
+            => DateTimeOffset.FromUnixTimeSeconds((Int32)value);
+        
         public override void SetValue(IDbDataParameter parameter, DateTimeOffset value)
             => parameter.Value = value;
